Move saml.server.ts into app/lib
