+++ app/routes/logout.tsx
@@ -2,7 +2,7 @@ import { redirect } from '@remix-run/node';
 import { getUsername } from '~/session.server';
 
 import type { ActionArgs } from '@remix-run/node';
-import { createLogoutRequest } from '~/saml.server';
+import { createLogoutRequest } from '~/lib/saml.server';
 
 export const action = async ({ request }: ActionArgs) => {
   const user = await getUsername(request);
